chore(node): refactor out replicate_data_batch code
